+++ ui/style.js
@@ -10,16 +10,17 @@ PSElement.prototype.style = function( property, value ){
   }
   
   
-  if( getStyleText ) return this.raw.style.cssText
+  if( propertyIsObject ) {
+    let keyPairs = Object.entries( property )
+    keyPairs.forEach( setStyleFromKeyPair )
+  }
+  
+  else if( getStyleText ) return this.raw.style.cssText
   
   else if( getStylePropertyValue ) return this.raw.style.getPropertyValue( property )
   
   else if( deleteStyleProperty ) this.raw.style.removeProperty( property )
   
-  if( propertyIsObject ) {
-    let keyPairs = Object.entries( property )
-    keyPairs.forEach( setStyleFromKeyPair )
-  }
   else {
     this.raw.style.setProperty( property, value)
   }