+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 PlayScriptJS Changelog
 ---
 
+### v0.3
+
+#### New
+- added a basic [splash screen](/core/splash-screen.md) with a progress bar, title text and subtext.
+- added default `charset` and `viewport` meta data at [enhancements](/core/enhancements.md).
+- added app title to [enhancements](/core/enhancements.md).
+- added a [theme builder](/core/theme.md) with default themes `light`, `dark` and `oled`.
+
+#### Fix
+- `PSElement.style` and `PSElement.attr` now supports setting multiple properties at once using objects ( broken in the previous version ).
+
+---
+
 ### v0.2
 
 #### New